connect component to redux
